# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/summary:
    get:
      operationId: getWalletSummary
      tags: [wallet]
      summary: Get the customer wallet summary
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Wallet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSummary"
        "401":
          description: Authentication required
  /wallet/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallet]
      summary: List wallet transactions for the signed-in user
      security: [{ clerkSession: [] }]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Wallet transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"
        "401":
          description: Authentication required
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Request a withdrawal of available balance
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWithdrawalRequest"
      responses:
        "201":
          description: Withdrawal request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
        "400":
          description: Invalid request or insufficient balance
        "401":
          description: Authentication required
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    WalletSummary:
      type: object
      properties:
        balance:
          type: number
        pendingCashback:
          type: number
        availableCashback:
          type: number
        rewardPoints:
          type: number
      required: [balance, pendingCashback, availableCashback, rewardPoints]
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        referenceId:
          type: ["string", "null"]
        referenceType:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, type, amount, description, referenceId, referenceType, createdAt]
    WithdrawalRequest:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        status:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, amount, status, bankName, accountNumber, notes, createdAt, updatedAt]
    CreateWithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        bankName:
          type: string
        accountNumber:
          type: string
        notes:
          type: string
      required: [amount, bankName, accountNumber]
...